+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * CLI version now shows engine as well
 * Added noScript option for clientPackages to exclude any main/mainMinified
+* HTML Template Script includes are generated at build time not configure
 
 ## v0.9.6